+++ sdk/identity/azure-identity/tests/test_application_credential.py
@@ -9,7 +9,7 @@ from urllib.parse import urlparse
 from azure.core.credentials import AccessToken, AccessTokenInfo
 from azure.identity import CredentialUnavailableError
 from azure.identity._credentials.application import AzureApplicationCredential
-from azure.identity._constants import EnvironmentVariables
+from azure.identity._constants import EnvironmentVariables, CLIENT_SECRET_VARS
 import pytest
 from urllib.parse import urlparse
 
@@ -30,7 +30,7 @@ def test_get_token(get_token_method):
             return mock_response(json_payload=build_aad_response(access_token=expected_token))
         return get_discovery_response("https://{}/{}".format(parsed.netloc, tenant_id))
 
-    with patch.dict("os.environ", {var: "..." for var in EnvironmentVariables.CLIENT_SECRET_VARS}, clear=True):
+    with patch.dict("os.environ", {var: "..." for var in CLIENT_SECRET_VARS}, clear=True):
         credential = AzureApplicationCredential(transport=Mock(send=send))
 
     token = getattr(credential, get_token_method)("scope")
@@ -97,7 +97,7 @@ def test_authority(authority):
                 assert "authority" not in kwargs
 
     # authority should be passed to EnvironmentCredential as a keyword argument
-    environment = {var: "foo" for var in EnvironmentVariables.CLIENT_SECRET_VARS}
+    environment = {var: "foo" for var in CLIENT_SECRET_VARS}
     with patch(AzureApplicationCredential.__module__ + ".EnvironmentCredential") as mock_credential:
         with patch.dict("os.environ", environment, clear=True):
             test_initialization(mock_credential, expect_argument=True)