+++ sdk/identity/azure-identity/tests/test_default_async.py
@@ -17,7 +17,7 @@ from azure.identity.aio import (
     SharedTokenCacheCredential,
     VisualStudioCodeCredential,
 )
-from azure.identity._constants import EnvironmentVariables
+from azure.identity._constants import EnvironmentVariables, CLIENT_SECRET_VARS
 import pytest
 
 from helpers import mock_response, Request, GET_TOKEN_METHODS
@@ -84,7 +84,7 @@ def test_authority(authority):
                 assert "authority" not in kwargs
 
     # authority should be passed to EnvironmentCredential as a keyword argument
-    environment = {var: "foo" for var in EnvironmentVariables.CLIENT_SECRET_VARS}
+    environment = {var: "foo" for var in CLIENT_SECRET_VARS}
     with patch(DefaultAzureCredential.__module__ + ".EnvironmentCredential") as mock_credential:
         with patch.dict("os.environ", environment, clear=True):
             test_initialization(mock_credential, expect_argument=True)